// src/app/application/attendance-system/system-management/staff-basic-info-management/model/staff-basic-info.model.ts
export class StaffBasicInfoModel {
    usrId?: string;
    usrName?: string;
    usrRealName?: string;
    usrSex?: string;
    usrAge?: string;
    usrPosition?: string;
    usrEmail?: string;
    usrMobile?: string;
    usrContent?: string;
    usrRegisterTime?: string;
    usrWorkStatus?: string;
    usrOffWorkTime?: string;
    usrMac?: string;
    refDepartCode?: string;
    partName?: string;

}